% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-chapter-df.R
\name{create_chapter_df}
\alias{create_chapter_df}
\title{Create a Chapter}
\usage{
create_chapter_df(pages)
}
\arguments{
\item{pages}{A data frame of pages.}
}
\value{
A data frame of pages with new columns.
}
\description{
Create a Chapter
}
