% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-chapters.R
\name{create_chapters}
\alias{create_chapters}
\title{Create Chapters}
\usage{
create_chapters(pages)
}
\arguments{
\item{pages}{A data frame of pages.}
}
\value{
A summarised data frame with chapters.
}
\description{
Create Chapters
}
