% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages-df.R
\name{pages_df}
\alias{pages_df}
\title{A Data Frame of Pages}
\usage{
pages_df(pagespath)
}
\arguments{
\item{pagespath}{Path to input pages.}
}
\value{
A data frame with columns ymd and text.
}
\description{
A Data Frame of Pages
}
