% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-pages.R
\name{read_pages}
\alias{read_pages}
\title{Read Pages from Disk}
\usage{
read_pages(path, years = c(2021))
}
\arguments{
\item{path}{Path to pages.}

\item{years}{Range of years to read.}
}
\value{
A data frame with columns ymd and text.
}
\description{
Read Pages from Disk
}
