% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise-chapter-df.R
\name{summarise_chapter_df}
\alias{summarise_chapter_df}
\title{Summarise Chapters}
\usage{
summarise_chapter_df(pages)
}
\arguments{
\item{pages}{A data frame of pages.}
}
\value{
A data frame of chapters.
}
\description{
Summarise Chapters
}
