% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pages-df.R
\name{pages_df}
\alias{pages_df}
\alias{read_pages}
\title{Read Morning Pages from Disk}
\usage{
pages_df(pagespath)

read_pages(pagespath, years = c(2021))
}
\arguments{
\item{pagespath}{Path to morning pages.}

\item{years}{Years to filter.}
}
\value{
A data frame with columns \code{ymd} and \code{text}.
}
\description{
Read Morning Pages from Disk
}
